Fix three type conversion warnings
Fixed NSInteger to NSUInteger conversion warning in PBGitIndexController.m:310
Fixed int to NSUInteger conversion warning in OpenRecentController.m:106
Fixed unnecessary cast in PBGitIndex.m:291

+++ Classes/Controllers/OpenRecentController.m
@@ -103,7 +103,7 @@
 		if(selectedResult != nil) {
 			int index = (int)[currentResults indexOfObject: selectedResult]-1;
 			if(index < 0) index = 0;
-			selectedResult = [currentResults objectAtIndex:index];
+			selectedResult = [currentResults objectAtIndex:(NSUInteger)index];
 			[resultViewer selectRowIndexes:[NSIndexSet indexSetWithIndex:index] byExtendingSelection:FALSE];
 			[resultViewer scrollRowToVisible:index];
 		}
